Moved NpmRange to Semver.Ranges namespace
